requirements.txt
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
Flask-Script==2.0.6
Flask==1.0.3
Jinja2==2.10.1
SQLAlchemy==1.3.4
Werkzeug==0.15.5
alembic==1.0.10
flasgger==0.9.2
itsdangerous==1.1.0
psycopg2==2.8.2
pytz==2019.1
requests==2.22.0
setproctitle==1.1.10
six==1.12.0
PyJWT==1.7.1
elasticsearch>=7.0.0,<8.0.0
marshmallow-sqlalchemy==0.27.0
pytest_elasticsearch==2.1.0